+++ src/Kiota.Builder/Refiners/PythonRefiner.cs
@@ -74,6 +74,7 @@ public class PythonRefiner : CommonLanguageRefiner, ILanguageRefiner
                     CodePropertyKind.Custom,
                     CodePropertyKind.QueryParameter,
                     CodePropertyKind.RequestBuilder,
+                    CodePropertyKind.BackingStore
                 },
                 static s => s.ToFirstCharacterLowerCase().ToSnakeCase());
             AddParentClassToErrorClasses(
@@ -179,7 +180,21 @@ public class PythonRefiner : CommonLanguageRefiner, ILanguageRefiner
 
             foreach (var param in m.Parameters)
             {
+                if (param.SerializationName != null)
+                    param.SerializationName = param.Name;
                 param.Name = param.Name.ToFirstCharacterLowerCase().ToSnakeCase();
+            }
+            if (parentClassM.IsOfKind(CodeClassKind.Model))
+            {
+                foreach (var prop in parentClassM.Properties)
+                {
+                    if (string.IsNullOrEmpty(prop.SerializationName))
+                    {
+                        prop.SerializationName = prop.Name;
+                    }
+
+                    parentClassM.RenameChildElement(prop.Name, prop.Name.ToFirstCharacterLowerCase().ToSnakeCase());
+                }
             }
         }
         else if (currentElement is CodeClass c)