+++ src/Kiota.Builder/Validation/UrlFormEncodedComplex.cs
@@ -25,7 +25,7 @@ public class UrlFormEncodedComplex : ValidationRule<OpenApiOperation>
         if (schema == null) return;
         if (!schema.IsObject())
             context.CreateWarning(nameof(UrlFormEncodedComplex), $"The operation {operationId} has a {schemaName} which is not an object type. This is not supported by Kiota and serialization will fail.");
-        if (schema.Properties.Any(static x => x.Value.IsObject() || x.Value.IsArray()))
+        if (schema.Properties.Any(static x => x.Value.IsObject()))
             context.CreateWarning(nameof(UrlFormEncodedComplex), $"The operation {operationId} has a {schemaName} with a complex properties and the url form encoded content type. This is not supported by Kiota and serialization of complex properties will fail.");
     }
 }